% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor.plot}
\alias{cor.plot}
\title{Make plot to investigate correlation structure}
\usage{
cor.plot(data)
}
\arguments{
\item{data}{data.frame with predictor variables to be considered}
}
\value{
A pretty overview plot
}
\description{
Make plot to investigate correlation structure
}
\examples{

# cor.plot of some data

cor.plot(obs.sel[,18:24])

}
\author{
Taken from ecospat package & modified by Philipp Brun
}
