% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dev.R
\name{exp_dev}
\alias{exp_dev}
\title{Get explained deviance from model fit}
\usage{
exp_dev(mod)
}
\arguments{
\item{mod}{a fitted model object (GLM or GAM)}
}
\value{
Estimate of explained deviance
}
\description{
Estimate explained deviance from model output
}
\examples{

# explained deviance for a glm
mod1=glm(X3740 ~ poly(bio_01,2) + poly(bio_01,2) + poly(forest_fraction,2),
data=obs.sel,family="binomial")
exp_dev(mod1)

# explained deviance for a gam
library(gam)
mod2=gam(X3740 ~ s(bio_01) + s(bio_01) + s(forest_fraction),
data=obs.sel,family="binomial")
exp_dev(mod2)

}
\author{
Taken from ecospat package & modified by Philipp Brun
}
