% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recstrat_regl.R
\name{recstrat_regl}
\alias{recstrat_regl}
\title{Random Ecologically Stratified Sampling of equal numbers}
\usage{
recstrat_regl(in_grid, sample_no)
}
\arguments{
\item{in_grid}{The stratification grid to be sampled.}

\item{sample_no}{The total number of pixels to be sampled.}
}
\value{
Returns a dataframe with the selected sampling locations their
coordinates and the strata they belong in.
}
\description{
This function randomly takes an equal number of samples per class
in the stratification layer.
}
\details{
The number of classes in the stratification layer is determined
automatically from the integer input map. If the number of pixels in a
class is higher than the number of samples, then a random selection without
re-substitution is performed, otherwise all pixels of that class are selected.
}
\examples{

bio3<- raster(system.file("external/bioclim/current/bio3.grd",package="biomod2"))
bio12<- raster(system.file("external/bioclim/current/bio12.grd",package="biomod2"))
B3.rcl<-rcls.grd(bio3,9)
B12.rcl<-rcls.grd(bio12,9)
B3B12.comb <- B12.rcl+B3.rcl*10
B3B12.regl_samples <- recstrat_prop(B3B12.comb,100)
plot(B3B12.comb)
points(B3B12.regl_samples$x,B3B12.regl_samples$y,pch=16,cex=0.6,col=B3B12.regl_samples$class)

}
\author{
Achilleas Psomas, Niklaus Zimmermann
}
